FEATURE: Apply design to forums
Generic theming for forums. Designed to pick up as much from the base
theme as possible, whilst still providing enough structural styling to
make the forums usable out of the box. Has been tested in Dew, Simple,
and SSExpress themes.

This is a major change in some ways, as previously the design component
of forums has been very minimal. However, without design the forums
module is one of the most confusing and inaccessible so I do think it
needs to be done. The changes within the templates themselves are
mostly around turning multiple tables, and divs styled like tables,
into one table, and moving a few of the elements from the header to the
footer.